+++ lib/src/websocket_browser/websocket_browser.dart
@@ -1,0 +1,6 @@
+import 'dart:html';
+
+String incrementerSocket(dynamic data, [WebSocket ws]) {
+  final int ret = int.parse(data) + 1;
+  return ret.toString();
+}