+++ lib/websocket.dart
@@ -1,0 +1,1 @@
+export 'src/websocket/websocket.dart';